VideoCommon: Allow BitfieldExtract in specialized shaders
